orchestrator/pyproject.toml
[tool.poetry]
name = "manoe-orchestrator"
version = "0.1.0"
description = "Multi-Agent Narrative Orchestration Engine - AI Orchestrator"
authors = ["IShalkin <mailtoshalkin@gmail.com>"]
readme = "README.md"
packages = [{include = "orchestrator", from = ".."}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.0"
openai = "^1.6.0"
anthropic = "^0.8.0"
google-generativeai = "^0.3.0"
redis = "^5.0.0"
qdrant-client = "^1.7.0"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
